Fix Decimal→String type cast crash and align trip completion between home and detail
- API: Change Decimal to float in AccommodationResponse.pricePerNight and
  ManualFlightResponse.price to prevent Pydantic v2 serializing them as
  JSON strings (caused 'String is not subtype of num?' on Flutter side)
- API: Add compute_completion_batch() to TripsService using 6-segment logic
  (dates, flights, accommodations, activities≥3, baggage≥5, budget>0) with
  4 batch SQL queries; expose completionPercentage on TripResponse
- Flutter: Fix TextButton theme minimumSize from Size.fromHeight(42) →
  Size(0, 42) to prevent infinite-width layout crash in Row/SliverList
- Flutter: Home completion now uses server-computed completionPercentage
  instead of local 5-field metadata check that showed false 100%
- Flutter: Align trip detail budget segment to use trip.budgetTotal > 0
  (same as server) instead of budgetSummary != null (always true)

+++ api/src/api/flights/manual/schemas.py
@@ -1,7 +1,6 @@
 """Schémas Pydantic pour les vols manuels."""
 
 from datetime import datetime
-from decimal import Decimal
 from uuid import UUID
 
 from pydantic import BaseModel, Field
@@ -16,7 +15,7 @@ class ManualFlightCreateRequest(BaseModel):
     arrivalAirport: str | None = None
     departureDate: datetime | None = None
     arrivalDate: datetime | None = None
-    price: Decimal | None = None
+    price: float | None = None
     currency: str | None = None
     notes: str | None = None
     flightType: str = "MAIN"
@@ -31,7 +30,7 @@ class ManualFlightUpdateRequest(BaseModel):
     arrivalAirport: str | None = None
     departureDate: datetime | None = None
     arrivalDate: datetime | None = None
-    price: Decimal | None = None
+    price: float | None = None
     currency: str | None = None
     notes: str | None = None
     flightType: str | None = None
@@ -48,7 +47,7 @@ class ManualFlightResponse(BaseModel):
     arrivalAirport: str | None = Field(None, alias="arrival_airport")
     departureDate: datetime | None = Field(None, alias="departure_date")
     arrivalDate: datetime | None = Field(None, alias="arrival_date")
-    price: Decimal | None = None
+    price: float | None = None
     currency: str | None = None
     notes: str | None = None
     flightType: str = Field(..., alias="flight_type")
